clean up of file structure

+++ mpc/src/main/java/de/antoniusstrauch/mpc/core/bridge/IBlindingFactorRepository.java
@@ -4,5 +4,5 @@ import java.util.UUID;
 
 public interface IBlindingFactorRepository {
 
-  Integer getBlindingFactor(UUID eventBatchId);
+  Long getBlindingFactor(UUID eventBatchId);
 }
