+++ mpc/src/main/java/de/antoniusstrauch/mpc/core/usecase/collector/MergeAttributeEvents.java
@@ -5,7 +5,6 @@ import de.antoniusstrauch.mpc.core.entity.AttributionResult;
 import de.antoniusstrauch.mpc.core.entity.EventBatch;
 import de.antoniusstrauch.mpc.core.entity.EventBatchPair;
 import lombok.extern.slf4j.Slf4j;
-import org.jetbrains.annotations.NotNull;
 
 @Slf4j
 public class MergeAttributeEvents extends AUsecase<EventBatchPair, AttributionResult> {
@@ -19,10 +18,10 @@ public class MergeAttributeEvents extends AUsecase<EventBatchPair, AttributionRe
   }
 
   @Override
-  public AttributionResult runUsecase( EventBatchPair eventBatchPair) {
+  public AttributionResult runUsecase(EventBatchPair eventBatchPair) {
     log.info("MergeAttributeEvents for eventBatchPair: " + eventBatchPair.getEventBatchPairId());
     EventBatch eventBatch = mergeBatches.runUsecase(eventBatchPair);
-    Integer result = attributeEvents.runUsecase(eventBatch);
+    Long result = attributeEvents.runUsecase(eventBatch);
     return AttributionResult.builder().result(result).build();
   }
 }
