+++ mpc/src/main/java/de/antoniusstrauch/mpc/core/usecase/helper/BlindEvents.java
@@ -3,7 +3,7 @@ package de.antoniusstrauch.mpc.core.usecase.helper;
 import de.antoniusstrauch.mpc.core.AUsecase;
 import de.antoniusstrauch.mpc.core.bridge.IBlindingFactorRepository;
 
-public class BlindEvents extends AUsecase<BlindEventsCommand, Integer> {
+public class BlindEvents extends AUsecase<BlindEventsCommand, Long> {
 
 
   private final IBlindingFactorRepository blindingFactorRepository;
@@ -13,8 +13,8 @@ public class BlindEvents extends AUsecase<BlindEventsCommand, Integer> {
   }
 
   @Override
-  public Integer runUsecase(BlindEventsCommand command) {
-    Integer blindingFactor = blindingFactorRepository.getBlindingFactor(
+  public Long runUsecase(BlindEventsCommand command) {
+    Long blindingFactor = blindingFactorRepository.getBlindingFactor(
         command.getEventBatchPairId());
 
     return command.getMatchKey() * blindingFactor;
